Moved fluxmatrix and wkinetics to other files. The main routines which once existed in their respective tools now exist as classes to be called by the functions.  The routines aren't yet general, which will need to be fixed, ultimately.
