+++ lib/west_tools/westpa/kinetics/__init__.py
@@ -14,5 +14,5 @@ from _kinetics import (calculate_labeled_fluxes, labeled_flux_to_rate, #@Unresol
                        nested_to_flat_matrix, nested_to_flat_vector, #@UnresolvedImport
                        flat_to_nested_matrix, flat_to_nested_vector, find_macrostate_transitions, #@UnresolvedImport
                        sequence_macro_flux_to_rate, sequence_macro_flux_to_rate_bs) #@UnresolvedImport
-
+from events import WKinetics
