+++ lib/west_tools/westpa/reweight/__init__.py
@@ -7,3 +7,4 @@ log = logging.getLogger(__name__)
 
 import _reweight
 from _reweight import (stats_process, reweight_for_c) #@UnresolvedImport
+from matrix import FluxMatrix
